requirements.txt
streamlit>=1.22.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.2.0
numpy>=1.23.0
matplotlib>=3.6.0
tqdm>=4.64.1  # For progress bars if used in your models